requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.1.0
nltk==3.8.1
# Removed transformers and torch to reduce Heroku slug size
# The app automatically falls back to lightweight_ai.py
numpy==1.24.4
pandas==2.1.3
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
requests==2.31.0